requirements.txt
python-telegram-bot==20.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
Pillow==10.1.0